requirement.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
audioop-lts==0.2.2
discord==2.3.2
discord.py==2.6.3
frozenlist==1.7.0
idna==3.10
multidict==6.6.4
pip==25.2
propcache==0.3.2
psycopg2==2.9.10
python-dotenv==1.1.1
yarl==1.20.1